try another favicon

+++ src/resolvers/UserMutation.js
@@ -1,17 +1,33 @@
 import * as R from 'ramda'
-import { getUserEmail, updateUserInAuth0, deleteUserInAuth0 } from '../utils/auth'
+import {
+	getUserEmail,
+	checkIsRoot,
+	updateUserInAuth0,
+	deleteUserInAuth0,
+} from '../utils/auth'
 
 const UserMutation = {
   async createUser(parent, args, { prisma, request }, info) {
-		return prisma.mutation.createUser(args, info)
+		return prisma.mutation.createUser({
+			data: {
+				...args.data,
+				role: 'USER', // overwrite role for protection
+			}
+		}, info)
 	},
 
 	async updateUser(parent, args, { prisma, request }, info) {
 		const email = await getUserEmail(prisma, request, args.where)
+		const isRoot = await checkIsRoot(prisma, request)
+
+		const data = {
+			...args.data,
+			role: args.data.role && isRoot ? args.data.role : 'USER'
+		}
 
 		// if user data to be updated includes 'email' or 'password', then update Auth0 as well since it involves auth
 		try {
-			const { email: newEmail, password: newPassword } = args.data
+			const { email: newEmail, password: newPassword } = data
 			if (newEmail || newPassword) {
 				const updates = R.mergeAll([newEmail ? { email: newEmail } : {}, newPassword ? { password: newPassword } : {}])
 				await updateUserInAuth0(email, updates)
@@ -24,9 +40,7 @@ const UserMutation = {
 			where: {
 				email
 			},
-			data: {
-				...args.data
-			}
+			data
 		}, info)
 	},
 
@@ -48,16 +62,25 @@ const UserMutation = {
   },
 
 	async updateManyUsers(parent, args, { prisma, request }, info) {
-		const email = await getUserEmail(prisma, request, args.where)
+		const isRoot = await checkIsRoot(prisma, request)
+
+		if (!isRoot) { throw new Error('You don\'t have permission to batch update users') }
 
 		// Don't allow batch updates for email or password
 		if (args.data.email || args.data.password) { throw new Error('Cannot batch update email or password') }
 
-		return prisma.mutation.updateManyUsers(args, info)
+		const data = {
+			...args.data,
+			role: 'USER', // overwrite role for protection
+		}
+
+		return prisma.mutation.updateManyUsers({ data }, info)
 	},
 
 	async deleteManyUsers(parent, args, { prisma, request }, info) {
-		const email = await getUserEmail(prisma, request, args.where)
+		const isRoot = await checkIsRoot(prisma, request)
+
+		if (!isRoot) { throw new Error('You don\'t have permission to batch delete users') }
 
 		return prisma.mutation.deleteManyUsers(args, info)
 	},
